Fixed a bug that caused validation on displaying previously-submitted answers a little overzealous.
